+++ examples/website/line/app.js
@@ -2,7 +2,8 @@ import React, {Component} from 'react';
 import {render} from 'react-dom';
 
 import {StaticMap} from 'react-map-gl';
-import DeckGL, {LineLayer, ScatterplotLayer} from 'deck.gl';
+import DeckGL from '@deck.gl/react';
+import {LineLayer, ScatterplotLayer} from '@deck.gl/layers';
 import GL from '@luma.gl/constants';
 
 // Set your mapbox token here