+++ modules/aggregation-layers/bundle.js
@@ -8,5 +8,7 @@ const deck = _global.deck || {};
 if (!deck.LineLayer) {
   throw new Error('@deck.gl/layers is not found');
 }
+// Merge experimental exports
+Object.assign(deckGLLayers.experimental, deck.experimental);
 
 module.exports = Object.assign(deck, deckGLLayers);