+++ modules/main/bundle.js
@@ -1,4 +1,5 @@
 const deck = require('../core/bundle');
+const {experimental} = deck;
 
 Object.assign(
   deck,
@@ -11,5 +12,8 @@ Object.assign(
   require('@deck.gl/mapbox'),
   require('@deck.gl/json')
 );
+
+// Make sure core exports are preserved
+Object.assign(deck.experimental, experimental);
 
 module.exports = deck;